applied Sebastian's recommended changes
