--- src/Squidex.Infrastructure.AzureStorage/IStorageAccountManager.cs
@@ -1,19 +1,0 @@
-﻿// ==========================================================================
-//  IStorageAccountManager.cs
-//  Squidex Headless CMS
-// ==========================================================================
-//  Copyright (c) Squidex Group
-//  All rights reserved.
-// ==========================================================================
-
-using Microsoft.WindowsAzure.Storage.Blob;
-
-namespace Squidex.Infrastructure.AzureStorage
-{
-    public interface IStorageAccountManager
-    {
-        CloudBlobClient CreateCloudBlobClient();
-
-        string GetSharedAccessSignature();
-    }
-}
